+++ rdt/__init__.py
@@ -5,19 +5,34 @@
 
 __author__ = """MIT Data To AI Lab"""
 __email__ = 'dailabmit@gmail.com'
-__version__ = '0.2.0'
+__version__ = '0.2.7.dev0'
 
 import numpy as np
 import pandas as pd
 
+from rdt import transformers
 from rdt.hyper_transformer import HyperTransformer
 
 __all__ = [
     'HyperTransformer',
+    'transformers'
 ]
 
 
 def get_demo(dtypes=('int', 'float', 'str', 'datetime'), nans=0.2, size=10):
+    """Generate random demo data with multiple data types.
+
+    Args:
+        dtypes (tuple or list):
+            Data types to include in the generated demo data. Defaults to all.
+        nans (float):
+            Proportion of null values to generate. Defaults to 0.2.
+        size (int):
+            Number of data rows to generate.
+
+    Returns:
+        pd.DataFrame
+    """
     if np.isscalar(nans):
         nans = [nans] * len(dtypes)